requirements.txt
# production
aiohttp==3.8.4
alembic==1.11.1
bcrypt==4.0.1
fastapi==0.99.0
httpx==0.24.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.6.0
SQLAlchemy==2.0.17
uvicorn==0.22.0

# development
black==23.3.0
flake8==6.0.0
mypy==1.4.1

# testing
coverage==7.2.7
pytest-asyncio==0.21.0
